+++ src/utils/errors/index.js
@@ -1,8 +1,9 @@
 export * from './errorCodes.js';
 export { default as BaseError } from './BaseError.js';
-export { default as InternalError } from './InternalError.js';
-export { default as UnknownError } from './UnknownError.js';
-export { default as SkippedError } from './SkippedError.js';
-export { default as InvalidConfigError } from './InvalidConfigError.js';
 export { default as EnvironmentError } from './EnvironmentError.js';
+export { default as InternalError } from './InternalError.js';
+export { default as InvalidConfigError } from './InvalidConfigError.js';
+export { default as InvalidInputError } from './InvalidInputError.js';
 export { default as ShellError } from './ShellError.js';
+export { default as SkippedError } from './SkippedError.js';
+export { default as UnknownError } from './UnknownError.js';